feat(sec-24): Add class192 function for sorting slices
Add a new function, class192, to demonstrate sorting slices in Go. This function uses the sort package to sort an integer slice and a string slice in ascending order. The purpose of this addition is to showcase the usage of the sort package and enhance understanding of sorting operations in Go.