feat(sec-26): Add class206 function for demonstrating atomic operations
This commit adds a new function, `class206`, to demonstrate the usage of atomic operations in Go programming. The function showcases the usage of the `sync/atomic` package to perform atomic operations on shared variables. It includes the creation of goroutines that increment a counter using `atomic.AddInt64` and print the current value using `atomic.LoadInt64`. The commit also includes necessary imports and prints information about the number of CPUs and goroutines before and after the goroutine creation. This addition enhances understanding of atomic operations and their role in concurrent programming.