feat(sec-27): Add class211 function for demonstrating atomic operations
This commit adds a new function, `class211`, to demonstrate the usage of atomic operations in Go programming. The function showcases the usage of the `sync/atomic` package to perform atomic operations on shared variables. It includes the creation of goroutines that increment a counter using `atomic.AddInt64` and print the current value using `atomic.LoadInt64`. This addition enhances understanding of atomic operations and their role in concurrent programming.