feat(sec-29): Add class224 function for demonstrating channel usage
This commit adds a new function, `class224`, to demonstrate the usage of channels in Go programming. The function generates a channel of integers using the `gen` function and receives and prints the values from the channel using the `receive` function. This addition enhances understanding of how to use channels for communication between goroutines and demonstrates the use of goroutines and channels for concurrent programming.