feat(sec-33): Add dog package to convert human years to dog years
This commit adds the `dog` package, which provides a function to convert human years to dog years. The `Years` function takes in the number of human years as an argument and returns the equivalent number of dog years by multiplying it by 7. This addition enhances the functionality of the codebase by providing a convenient way to calculate dog years.