feat(sec-34): Add Sum function to package acdc wit test example
This commit adds the `Sum` function to the `acdc` package in the `main.go` file. The `Sum` function takes an unlimited number of values of type `int` and returns their sum. This addition enhances the functionality of the codebase by providing a convenient way to calculate the sum of multiple numbers.